// src/lib/api/generated/model/templateTypeEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `WELCOME` - Welcome
* `REMINDER` - Reminder
* `FOLLOWUP` - Follow-up
* `THANK_YOU` - Thank You
 */
export type TemplateTypeEnum = typeof TemplateTypeEnum[keyof typeof TemplateTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TemplateTypeEnum = {
  WELCOME: 'WELCOME',
  REMINDER: 'REMINDER',
  FOLLOWUP: 'FOLLOWUP',
  THANK_YOU: 'THANK_YOU',
} as const;